requirements-sqlite.txt
aiogram==3.4.1
sqlalchemy==2.0.23
aiosqlite
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
jdatetime==5.1.0
aiohttp==3.9.2
alembic==1.13.1
python-multipart==0.0.6
cryptography==41.0.7
